build_kernel.sh: Bump simultaneously running jobs to 32
This commit bumps the simultaneously running jobs (specified by the -j argument) to 32 to allow more jobs to run at a time which should speed up the builds on GitHub Actions. If you're building from a low end computer, it is recommended to lower this amount locally or use a value corresponding to the amount of cores and threads in your computer.